pyproject.toml
[project]
name = "infraassistant"
version = "0.1.0"
description = "InfraAssistant is a tool to help manage infrastructure resources efficiently"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=1.3.2",
    "fastapi>=0.115.6",
    "langsmith>=0.2.6",
    "litellm>=1.55.10",
    "openai>=1.58.1",
    "pydantic==2.10.1",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ipython>=8.31.0",
    "pytest>=8.3.4",
    "ruff>=0.8.4",
]
